refactor: dateFormatter ViewModel 로 이동 #3
